- add lombok and embeddable to Address
